+++ addons/docs/src/blocks/Description.tsx
@@ -31,7 +31,7 @@ const getDocgen = (component?: Component) =>
 
 export const getDescriptionProps = (
   { of, type, markdown }: DescriptionProps,
-  { parameters, getPropDefs }: DocsContextProps
+  { parameters }: DocsContextProps
 ): PureDescriptionProps => {
   if (markdown) {
     return { markdown };