+++ addons/docs/src/blocks/Preview.tsx
@@ -4,6 +4,7 @@ import {
   Preview as PurePreview,
   PreviewProps as PurePreviewProps,
 } from '@storybook/components';
+import { toId } from '@storybook/router';
 import { Story } from './Story';
 import { getSourceProps } from './Source';
 import { DocsContext, DocsContextProps } from './DocsContext';
@@ -24,18 +25,17 @@ const getPreviewProps = (
     children,
     ...props
   }: PreviewProps & { children?: React.ReactNode },
-  { storyStore }: DocsContextProps
+  { mdxKind, storyStore }: DocsContextProps
 ): PurePreviewProps => {
-  const extraProps = {};
   if (withSource === SourceState.NONE && !children) {
     return props;
   }
   const childArray: ReactNodeArray = Array.isArray(children) ? children : [children];
-  const childWithId: React.ReactElement = childArray.find(
-    (c: React.ReactElement) => c.props && c.props.id
-  ) as ReactElement;
-  const targetId = childWithId ? childWithId.props.id : null;
-  const sourceProps = getSourceProps({ id: targetId }, { storyStore });
+  const stories = childArray.filter(
+    (c: React.ReactElement) => c.props && (c.props.id || c.props.name)
+  ) as React.ReactElement[];
+  const targetIds = stories.map(s => s.props.id || toId(mdxKind, s.props.name));
+  const sourceProps = getSourceProps({ ids: targetIds }, { storyStore });
   return {
     ...props, // pass through columns etc.
     withSource: sourceProps,