+++ addons/docs/src/blocks/Props.tsx
@@ -1,16 +1,27 @@
 import React from 'react';
-import { PropsTable, PropsTableError, PropsTableProps } from '@storybook/components';
+import { PropsTable, PropsTableError, PropsTableProps, PropDef } from '@storybook/components';
 import { DocsContext, DocsContextProps } from './DocsContext';
 import { CURRENT_SELECTION } from './shared';
+import { getPropDefs as autoPropDefs, PropDefGetter } from '../lib/getPropDefs';
 
 interface PropsProps {
   exclude?: string[];
   of: any;
 }
 
+const inferPropDefs = (framework: string): PropDefGetter | null => {
+  switch (framework) {
+    case 'react':
+    case 'vue':
+      return autoPropDefs;
+    default:
+      return null;
+  }
+};
+
 export const getPropsTableProps = (
   { exclude, of }: PropsProps,
-  { parameters, getPropDefs }: DocsContextProps
+  { parameters }: DocsContextProps
 ): PropsTableProps => {
   const { component } = parameters;
   try {
@@ -18,11 +29,16 @@ export const getPropsTableProps = (
     if (!target) {
       throw new Error(PropsTableError.NO_COMPONENT);
     }
+
+    const { framework = null } = parameters || {};
+    const { getPropDefs = inferPropDefs(framework) } =
+      (parameters && parameters.options && parameters.options.docs) || {};
+
     if (!getPropDefs) {
       throw new Error(PropsTableError.PROPS_UNSUPPORTED);
     }
     const allRows = getPropDefs(target);
-    const rows = !exclude ? allRows : allRows.filter(row => !exclude.includes(row.name));
+    const rows = !exclude ? allRows : allRows.filter((row: PropDef) => !exclude.includes(row.name));
     return { rows };
   } catch (err) {
     return { error: err.message };