+++ examples/cra-mdx/.storybook/webpack.config.js
@@ -2,10 +2,10 @@ const path = require('path');
 
 module.exports = async ({ config }) => {
   config.module.rules.push({
-    test: /\.stories\.js$/,
+    test: /\.stories\.[tj]sx?$/,
     use: [
       {
-        loader: require.resolve('@storybook/addon-storysource/loader'),
+        loader: require.resolve('@storybook/source-loader'),
         options: { injectParameters: true },
       },
     ],