+++ examples/cra-mdx/src/stories/module.stories.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import TypescriptButton from './TypescriptButton';
+
+export const componentMeta: any = {
+  title: 'Typescript|Module',
+  decorators: [],
+  parameters: { component: TypescriptButton },
+};
+
+export const small = (): any => <TypescriptButton variant="small" />;
+export const large = (): any => <TypescriptButton variant="large" />;