+++ examples/vue-kitchen-sink/.storybook/config.js
@@ -2,6 +2,7 @@ import { load, addParameters, addDecorator } from '@storybook/vue';
 import Vue from 'vue';
 import Vuex from 'vuex';
 import { withA11y } from '@storybook/addon-a11y';
+import { DocsPage } from '@storybook/addon-docs/blocks';
 import MyButton from '../src/stories/Button.vue';
 
 addDecorator(withA11y);
@@ -11,7 +12,11 @@ Vue.use(Vuex);
 addParameters({
   options: {
     hierarchyRootSeparator: /\|/,
+    docs: {
+      iframeHeight: '60px',
+    },
   },
+  docs: DocsPage,
 });
 
 load(require.context('../src', true, /\.stories\.js$/), module);