+++ examples/vue-kitchen-sink/src/stories/index.stories.js
@@ -2,6 +2,7 @@ import { storiesOf } from '@storybook/vue';
 import { linkTo } from '@storybook/addon-links';
 
 import Welcome from './Welcome.vue';
+import Button from './Button.vue';
 import App from '../App.vue';
 
 storiesOf('Welcome', module).add('Welcome', () => ({
@@ -13,6 +14,7 @@ storiesOf('App', module).add('App', () => ({
 }));
 
 storiesOf('Button', module)
+  .addParameters({ component: Button })
   // Works if Vue.component is called in the config.js in .storybook
   .add('rounded', () => ({
     template: '<my-button :rounded="true">A Button with rounded edges</my-button>',